Extract comments from the Ticket page
