+++ Issue Tracker/frontend/src/pages/TicketPage.tsx
@@ -2,19 +2,11 @@ import axios from 'axios';
 import { useState, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import '../styles/TicketPage.css';
+import Comments from '../components/Comments';
 
 interface User {
   id: string;
   username: string;
-}
-
-interface Comment {
-  id: string;
-  userId: string;
-  ticketId: string;
-  content: string;
-  createdBy: string;
-  createdAt: string;
 }
 
 type ErrorType = { [key: string]: string };
@@ -30,11 +22,6 @@ const TicketPage = () => {
   const [loading, setLoading] = useState(true);
   const [fetchError, setFetchError] = useState('');
   const [hasChanges, setHasChanges] = useState(false);
-  const [comments, setComments] = useState<Comment[]>([]);
-  const [newComment, setNewComment] = useState('');
-  const [commentError, setCommentError] = useState('');
-  const [addingComment, setAddingComment] = useState(false);
-  const [showAddCommentForm, setShowAddCommentForm] = useState(false);
 
   useEffect(() => {
     axios.get(`/api/tickets/${id}`)
@@ -53,19 +40,6 @@ const TicketPage = () => {
       })
       .catch(() => setFetchError('Failed to load users'));
   }, []);
-
-  useEffect(() => {
-    if (id) {
-      axios.get(`/api/tickets/${id}/comments`)
-        .then(({ data }) => {
-          setComments(data);
-          console.log('Comments loaded:', data);
-        })
-        .catch(() => {
-          console.error('Failed to load comments');
-        });
-    }
-  }, [id]);
 
   useEffect(() => {
     if (ticket && originalTicket) {
@@ -299,51 +273,6 @@ const TicketPage = () => {
     }).format(new Date(isoString));
   };
 
-  const handleCommentChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setNewComment(e.target.value);
-    if (e.target.value.trim().length < 1) {
-      setCommentError('Comment cannot be empty');
-    } else {
-      setCommentError('');
-    }
-  };
-
-  const handleAddComment = (e: React.FormEvent) => {
-    e.preventDefault();
-    
-    if (!newComment.trim()) {
-      setCommentError('Comment cannot be empty');
-      return;
-    }
-
-    setAddingComment(true);
-    
-    axios.post(`/api/tickets/${id}/comments`, { 
-      content: newComment,
-      ticketId: id
-    })
-      .then(({ data }) => {
-        setComments([...comments, data]);
-        setNewComment('');
-        setCommentError('');
-        setShowAddCommentForm(false);
-      })
-      .catch(error => {
-        setCommentError(`Failed to add comment: ${error.message}`);
-      })
-      .finally(() => {
-        setAddingComment(false);
-      });
-  };
-
-  const toggleAddCommentForm = () => {
-    setShowAddCommentForm(!showAddCommentForm);
-    if (!showAddCommentForm) {
-      setNewComment('');
-      setCommentError('');
-    }
-  };
-
   return (
     <div className="ticket-container">
       <h1 className="ticket-title">
@@ -407,69 +336,7 @@ const TicketPage = () => {
         </div>
       )}
 
-      <div className="comments-section">
-        <div className="comments-header">
-          <h2 className="comments-title">
-            Comments ({comments.length})
-          </h2>
-        </div>
-        
-        <div className="comments-content">
-          {comments.length === 0 ? (
-            <p className="no-comments">No comments yet</p>
-          ) : (
-            <div className="comments-list">
-              {comments.map((comment) => (
-                <div key={comment.id} className="comment">
-                  <div className="comment-header">
-                    <span className="comment-author">Posted by: {getAssigneeName(comment.createdBy)}</span>
-                    <span className="comment-date">Create date: {formatDate(comment.createdAt)}</span>
-                  </div>
-                  <div className="comment-content">{comment.content}</div>
-                </div>
-              ))}
-            </div>
-          )}
-          
-          {!showAddCommentForm ? (
-            <button 
-              onClick={toggleAddCommentForm} 
-              className="btn-secondary add-comment-button"
-            >
-              Add Comment
-            </button>
-          ) : (
-            <div className="add-comment">
-              <form onSubmit={handleAddComment}>
-                <textarea
-                  value={newComment}
-                  onChange={handleCommentChange}
-                  className={`comment-textarea ${commentError ? 'input-error' : ''}`}
-                  placeholder="Write your comment here..."
-                  rows={4}
-                ></textarea>
-                {commentError && <div className="error-message">{commentError}</div>}
-                <div className="comment-form-actions">
-                  <button 
-                    type="submit" 
-                    className="btn-primary" 
-                    disabled={addingComment || !newComment.trim()}
-                  >
-                    {addingComment ? 'Adding...' : 'Submit'}
-                  </button>
-                  <button 
-                    type="button" 
-                    className="btn-secondary"
-                    onClick={toggleAddCommentForm}
-                  >
-                    Cancel
-                  </button>
-                </div>
-              </form>
-            </div>
-          )}
-        </div>
-      </div>
+      <Comments ticketId={id!} users={users} />
     </div>
   );
 };
